requirements.txt
astropy-helpers
sphinx-astropy
sphinx
pip==18.1
setuptools==40.8.0
gitpython
astropy
numpy
glue-core
pyqt5
specviz
reproject
